backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-auth==2.23.4
Pillow==10.1.0
django-storages==1.14.2
django-filter==23.4
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
python-jose==3.3.0
requests==2.31.0 